Weekly and fortnightly cut-offs, for lending that does not run on months
Buy-now-pay-later is four fortnightly payments. Payday lending runs against a
pay cycle. Group-collected microfinance is weekly. On the nearest cadence the
engine had - daily - a four-instalment plan carries 84 rows instead of 4, twenty
times the data describing the same product.

Three places: the freq literal, the periods-per-year map, and the offset table.
A fortnight is two anchored weeks rather than a separate offset, because pandas
has none.

periods_per_year is 365.25/7 and 365.25/14, not 52 and 26. Rounding would bias
every annual-to-period rate conversion in one direction and compound it across
the panel - small, invisible, and always the same way.

Also fixes a circular import this uncovered: calendar -> spec -> loader ->
generate -> book -> calendar. Nothing had hit it because every caller happened
to import sdd.generate first, so the loop was already broken by the time
calendar was reached, but 'from sdd.calendar import period_dates' in a fresh
interpreter failed outright. Calendar is only a type annotation there, so it
moves under TYPE_CHECKING.

Verified end to end on a BNPL product: 3,000 plans over 8 fortnights, four
instalments each, 9/9 invariants, cut-offs 14 days apart.

+++ src/sdd/generate/groups.py
@@ -182,7 +182,7 @@ def attach_groups(
         else:
             # A cohort joining an open pool: some entities belong to borrowers
             # the book already has, the rest bring new ones.
-            new_count = int(round(n * group.new_group_rate))
+            new_count = round(n * group.new_group_rate)
             minted = pd.DataFrame()
             if new_count:
                 wanted = max(1, group.group_count(new_count))
